create rmf_websocket pkg
