+++ rmf_task_ros2/CMakeLists.txt
@@ -18,11 +18,11 @@ find_package(rmf_api_msgs REQUIRED)
 find_package(rmf_traffic REQUIRED)
 find_package(rmf_traffic_ros2 REQUIRED)
 find_package(rmf_task_msgs REQUIRED)
+find_package(rmf_websocket REQUIRED)
 find_package(Eigen3 REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(nlohmann_json REQUIRED)
 find_package(nlohmann_json_schema_validator_vendor REQUIRED)
-find_package(websocketpp REQUIRED)
 
 file(GLOB_RECURSE core_lib_srcs "src/rmf_task_ros2/*.cpp")
 add_library(rmf_task_ros2 SHARED ${core_lib_srcs})
@@ -32,9 +32,9 @@ target_link_libraries(rmf_task_ros2
     rmf_api_msgs::rmf_api_msgs
     rmf_traffic::rmf_traffic
     rmf_traffic_ros2::rmf_traffic_ros2
+    rmf_websocket::rmf_websocket
     ${rmf_task_msgs_LIBRARIES}
     ${rclcpp_LIBRARIES}
-    ${websocketpp_LIBRARIES}
     nlohmann_json::nlohmann_json
     nlohmann_json_schema_validator
 )
@@ -45,8 +45,8 @@ target_include_directories(rmf_task_ros2
     $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
     ${rmf_traffic_ros2_INCLUDE_DIRS}
     ${rmf_task_msgs_INCLUDE_DIRS}
+    ${rmf_websocket_INCLUDE_DIRS}
     ${rclcpp_INCLUDE_DIRS}
-    ${WEBSOCKETPP_INCLUDE_DIR}
 )
 
 ament_export_targets(rmf_task_ros2 HAS_LIBRARY_TARGET)
@@ -76,6 +76,7 @@ if(BUILD_TESTING)
   target_link_libraries(test_rmf_task_ros2
       rmf_task_ros2
       rmf_traffic::rmf_traffic
+      rmf_websocket::rmf_websocket
       rmf_traffic_ros2::rmf_traffic_ros2
       -pthread
   )
