+++ rmf_websocket/README.md
@@ -1,0 +1,7 @@
+# rmf\_websocket package
+
+This package provides a websocker wrapper client library to interact with websocket server.
+
+## Quality Declaration
+
+This package claims to be in the **Quality Level 4** category. See the [Quality Declaration](QUALITY_DECLARATION.md) for more details.
