+++ rmf_websocket/include/rmf_websocket/BroadcastClient.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 Open Source Robotics Foundation
+ * Copyright (C) 2022 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,15 +15,13 @@
  *
 */
 
-#ifndef SRC__RMF_TASK_ROS2__BROADCASTCLIENT_HPP
-#define SRC__RMF_TASK_ROS2__BROADCASTCLIENT_HPP
-
-#include <websocketpp/config/asio_no_tls_client.hpp>
-#include <websocketpp/client.hpp>
+#ifndef SRC__RMF_WEBSOCKET__BROADCASTCLIENT_HPP
+#define SRC__RMF_WEBSOCKET__BROADCASTCLIENT_HPP
 
 #include <nlohmann/json.hpp>
 
 #include <rclcpp/node.hpp>
+#include <rmf_utils/impl_ptr.hpp>
 
 #include <memory>
 #include <set>
@@ -32,30 +30,28 @@
 #include <thread>
 #include <atomic>
 
-namespace rmf_task_ros2 {
+namespace rmf_websocket {
 //==============================================================================
 // A wrapper around a websocket client for broadcasting states and logs for
 // fleets, robots and tasks. A queue of json msgs is maintained and published
 // in an internal thread whenever connection to a server is established.
-//
-// \Note: this class is a "copy" of rmf_fleet_adapter::BroadcastClient
-// TODO(YL) create a common broadcastclient lib
 class BroadcastClient : public std::enable_shared_from_this<BroadcastClient>
 {
 public:
-  using WebsocketClient =
-    websocketpp::client<websocketpp::config::asio_client>;
-  using WebsocketMessagePtr = WebsocketClient::message_ptr;
-  using ConnectionHDL = websocketpp::connection_hdl;
-  using Connections = std::set<ConnectionHDL, std::owner_less<ConnectionHDL>>;
+  using ProvideJsonUpdates =  std::function<std::vector<nlohmann::json>()>;
 
   /// \param[in] uri
   ///   "ws://localhost:9000"
   ///
   /// \param[in] node
+  ///
+  /// \param[in] on_open_connection_fn
+  ///   this function callback will be called when ws client
+  ///   is connected to the server
   static std::shared_ptr<BroadcastClient> make(
     const std::string& uri,
-    const std::shared_ptr<rclcpp::Node>& node);
+    const std::shared_ptr<rclcpp::Node>& node,
+    ProvideJsonUpdates on_open_connection_fn = nullptr);
 
   // Publish a single message
   void publish(const nlohmann::json& msg);
@@ -63,24 +59,13 @@ public:
   // Publish a vector of messages
   void publish(const std::vector<nlohmann::json>& msgs);
 
-  ~BroadcastClient();
+  class Implementation;
 
 private:
   BroadcastClient();
-  std::string _uri;
-  std::shared_ptr<rclcpp::Node> _node;
-  WebsocketClient _client;
-  websocketpp::connection_hdl _hdl;
-  std::mutex _wait_mutex;
-  std::mutex _queue_mutex;
-  std::condition_variable _cv;
-  std::queue<nlohmann::json> _queue;
-  std::thread _processing_thread;
-  std::thread _client_thread;
-  std::atomic_bool _connected;
-  std::atomic_bool _shutdown;
+  rmf_utils::unique_impl_ptr<Implementation> _pimpl;
 };
 
-} // namespace rmf_task_ros2
+} // namespace rmf_websocket
 
-#endif // SRC__RMF_TASK_ROS2__BROADCASTCLIENT_HPP
+#endif // SRC__RMF_WEBSOCKET__BROADCASTCLIENT_HPP
