+++ rmf_websocket/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>rmf_task_ros2</name>
+  <name>rmf_websocket</name>
   <version>2.0.0</version>
-  <description>A package managing the dispatching of tasks in RMF system.</description>
+  <description>A package managing the websocket api endpoints in RMF system.</description>
   <maintainer email="youliang@openrobotics.org">youliang</maintainer>
   <maintainer email="marco@openrobotics.org">Marco A. Gutiérrez</maintainer>
   <license>Apache License 2.0</license>
@@ -11,12 +11,9 @@
   <buildtool_depend>ament_cmake</buildtool_depend>
 
   <depend>rmf_utils</depend>
-  <depend>rmf_api_msgs</depend>
-  <depend>rmf_traffic</depend>
-  <depend>rmf_traffic_ros2</depend>
-  <depend>rmf_task_msgs</depend>
   <depend>rclcpp</depend>
-  <depend>rmf_websocket</depend>
+
+  <depend>libwebsocketpp-dev</depend>
   <depend>nlohmann-json-dev</depend>
   <depend>nlohmann_json_schema_validator_vendor</depend>
